# docs/swagger.yaml
definitions:
  dto.AdminInfoOutput:
    properties:
      avatar:
        type: string
      id:
        type: integer
      introduction:
        type: string
      login_time:
        type: string
      name:
        type: string
      roles:
        items:
          type: string
        type: array
    type: object
  dto.AdminLoginInput:
    properties:
      password:
        description: 密码
        example: "123456"
        type: string
      username:
        description: 管理员用户名
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  dto.AdminLoginOutput:
    properties:
      token:
        description: token
        example: token
        type: string
    type: object
  dto.ChangePwdInput:
    properties:
      password:
        description: 密码
        example: "123456"
        type: string
    required:
    - password
    type: object
  middleware.Response:
    properties:
      data:
        type: object
      errmsg:
        type: string
      errno:
        type: integer
      stack:
        type: object
      trace_id:
        type: object
    type: object
info:
  contact: {}
paths:
  /admin/admin_info:
    get:
      consumes:
      - application/json
      description: 管理员信息
      operationId: /admin/admin_info
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminInfoOutput'
              type: object
      summary: 管理员信息
      tags:
      - 管理员接口
  /admin/change_pwd:
    post:
      consumes:
      - application/json
      description: 修改密码
      operationId: /admin/change_pwd
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePwdInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 修改密码
      tags:
      - 管理员接口
  /admin_login/login:
    post:
      consumes:
      - application/json
      description: 管理员登陆
      operationId: /admin_login/login
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AdminLoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminLoginOutput'
              type: object
      summary: 管理员登陆
      tags:
      - 管理员接口
  /admin_login/logout:
    get:
      consumes:
      - application/json
      description: 管理员退出
      operationId: /admin_login/logout
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 管理员退出
      tags:
      - 管理员接口
swagger: "2.0"